api-gateway/requirements_conference.txt
# Intelligent API Gateway - Conference Paper Implementation
# Complete Python Requirements
# Python 3.11+ required

# ====================
# Web Framework & Server
# ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# ====================
# Authentication & Security
# ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0

# ====================
# Database & Caching (Redis + Supabase)
# ====================
redis==5.0.1
supabase==1.0.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# ====================
# Machine Learning (Paper Requirements)
# ====================
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2
scipy==1.11.3

# ====================
# HTTP Clients & Async
# ====================
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.1
requests==2.31.0

# ====================
# Data Processing & Validation
# ====================
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# ====================
# Monitoring & Logging
# ====================
structlog==23.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0
psutil==5.9.6

# ====================
# WebSocket Support
# ====================
websockets==12.0
python-socketio==5.10.0

# ====================
# Date/Time & Utilities
# ====================
python-dateutil==2.8.2
pytz==2023.3.post1
typing-extensions==4.8.0

# ====================
# Environment & Configuration
# ====================
python-dotenv==1.0.0
pyyaml==6.0.1

# ====================
# Testing (Development)
# ====================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# ====================
# Development Tools
# ====================
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# ====================
# Documentation
# ====================
mkdocs==1.5.3
mkdocs-material==9.4.8

# ====================
# Production Deployment
# ====================
gunicorn==21.2.0

# ====================
# Note: Install with:
# pip install -r requirements.txt
# 
# For development with all tools:
# pip install -r requirements.txt
#
# For production (minimal):
# pip install fastapi uvicorn redis supabase scikit-learn httpx python-jose passlib python-dotenv
# ====================
